+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/Models/Unity/MaterialWrapperBuilder.cs
@@ -19,7 +19,7 @@ namespace KRT.VRCQuestTools.Models.Unity
         {
 #pragma warning disable SA1136 // Enum values should be on separate lines
 #pragma warning disable SA1602 // Enumeration items should be documented
-            UTS2, Arktoon, Standard, Unlit, Quest, Sunao, AXCS, LilToon, ExtraSupport, Unverified,
+            UTS2, Arktoon, Standard, Unlit, Quest, Sunao, AXCS, LilToon, Poiyomi, ExtraSupport, Unverified,
 #pragma warning restore SA1602 // Enumeration items should be documented
 #pragma warning restore SA1136 // Enum values should be on separate lines
         }
@@ -43,6 +43,8 @@ namespace KRT.VRCQuestTools.Models.Unity
                     return new SunaoMaterial(material);
                 case ShaderCategory.LilToon:
                     return new LilToonMaterial(material);
+                case ShaderCategory.Poiyomi:
+                    return new PoiyomiMaterial(material);
                 case ShaderCategory.Standard:
                 case ShaderCategory.Unlit:
                 case ShaderCategory.Quest:
@@ -95,6 +97,10 @@ namespace KRT.VRCQuestTools.Models.Unity
             if (shaderName.Contains("liltoon"))
             {
                 return ShaderCategory.LilToon;
+            }
+            if (shaderName.Contains("poiyomi"))
+            {
+                return ShaderCategory.Poiyomi;
             }
             return ShaderCategory.Unverified;
         }